Rename app.py to better indicate its purpose.
Initially I had intended app.py to be a Flask app that did everything.
However, it makes more sense to separate this out as an importable
module that a Flask app (or delayed job worker, whatever...) can employ
to run arbitrary MATLAB code. In this vein it makes more sense to type
  from MATLABRunner import *
than
  from app import *